fix: PR #418 R6 — types 분리 + hook 선언 순서 예외 WHY 주석 + test mock reset
R6 review (Suggestion 4건, Blocker 0):
- S1: useChat.ts에 useSymbolChat()을 useState/useRef보다 먼저 호출하는 이유 WHY 주석 추가 (storageKeyRef 초기값이 timeframeFromCtx에 의존).
- S2: 4개 컴포넌트(ChartContent, FundamentalAiSummary, NewsAiSummary, OverallContent)에 usePublishSymbolChat이 useMemo 뒤에 위치하는 구조적 예외 WHY 주석 추가.
- S3: ChatPanel.test.tsx에 beforeEach로 mockIsAnalysisReady 명시적 초기화 — 테스트 간 모듈 스코프 mock state 누출 방지.
- S4: OverallAnalysisState 타입을 components/overall/types.ts로 분리 — utils/buildChatState.ts가 hooks/useOverallAnalysis.ts에 의존하던 역방향 의존 해소. hooks와 utils 양쪽이 동일 위치에서 import.
