+++ src/components/fundamental/FundamentalAiSummary.tsx
@@ -128,6 +128,8 @@ export function FundamentalAiSummary({ symbol }: FundamentalAiSummaryProps) {
     // with this page's numbers (not stale chart context). `timeframe` is null —
     // fundamental analysis is timeframe-agnostic; the launcher falls back to
     // DEFAULT_TIMEFRAME for the chat request.
+    // 훅 선언 순서 예외(MISTAKES.md #17): usePublishSymbolChat은 chatState(파생 변수)를
+    // 인자로 받기 때문에 useMemo 뒤에 위치해야 한다.
     const chatState = useMemo(
         () => ({
             context: { kind: 'fundamental', payload: result } as const,
