+++ src/components/overall/OverallContent.tsx
@@ -30,6 +30,8 @@ export function OverallContent({ symbol, timeframe }: OverallContentProps) {
     const modelId = useDefaultModelId();
     const { state, trigger } = useOverallAnalysis(symbol, timeframe, modelId);
 
+    // 훅 선언 순서 예외(MISTAKES.md #17): usePublishSymbolChat은 chatState(파생 변수)를
+    // 인자로 받기 때문에 useMemo 뒤에 위치해야 한다.
     const chatState = useMemo(
         () => buildChatState(state, timeframe),
         [state, timeframe]
